move KafkaFilter interface to filters package
